+++ components/motor/esp_simplefoc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ChangeLog
 
+## v1.2.0 - 2024-12-31
+### Enhancements:
+* FOC:
+  *  The angle sensor initializes I2C using the ``i2c_bus`` component to support both new and old I2C driver.
+
+### Bug Fix:
+* FOC:
+  *  Fix the issue where the pins used by the SPI sensor in TEST_APPS exceed the pin range of the ESP32.
+
 ## v1.1.0 - 2024-11-26
 ### Enhancements:
 * FOC: